Add cores optimization

+++ src/main/java/fr/telecompt/shavadoop/util/LocalRepoFile.java
@@ -1,9 +1,13 @@
 package fr.telecompt.shavadoop.util;
 
+import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.util.List;
 import java.util.Map;
@@ -51,22 +55,35 @@ public class LocalRepoFile {
 	}
 	
 	public static void writeFileFromPair(String nameFile, List<Pair> content) {
-		try {
-			FileWriter fw = new FileWriter (nameFile);
-			BufferedWriter bw = new BufferedWriter (fw);
-			PrintWriter write = new PrintWriter (bw); 
-			
-			for(Pair p : content) {
-			    write.println (p.getKey()
-			    		+ Constant.FILE_SEPARATOR
-			    		+ p.getValue()); 
+      	 // If the file exist, we concat
+      	 if(new File(nameFile).exists()) {
+      		try(
+      			PrintWriter out = new PrintWriter(new BufferedWriter(new FileWriter(nameFile, true)))) {
+				for(Pair p : content) {
+					out.println (p.getKey()
+				    		+ Constant.FILE_SEPARATOR
+				    		+ p.getValue()); 
+				}
+      		}catch (Exception e) {e.printStackTrace();}
+      	 // we create the file
+      	 } else {
+			try {
+				FileWriter fw = new FileWriter (nameFile);
+				BufferedWriter bw = new BufferedWriter (fw);
+				PrintWriter write = new PrintWriter (bw); 
+				
+				for(Pair p : content) {
+				    write.println (p.getKey()
+				    		+ Constant.FILE_SEPARATOR
+				    		+ p.getValue()); 
+				}
+				
+				write.close();
 			}
-			
-			write.close();
-		}
-		catch (Exception e){
-			System.out.println(e.toString());
-		}
+			catch (Exception e){
+				System.out.println(e.toString());
+			}
+      	 }
 	}
 	
 	public static void createDirectory(File file) {
